requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.7
django-environ==0.11.2
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0